Update README and refactoring some packages/classes
