+++ SRStudyProject/src/main/java/utils/KeysURL.java
@@ -1,4 +1,4 @@
-package core;
+package utils;
 
 public class KeysURL {
